Compute the block size on the context and change HMAC api.
compute the block on the contextinstead of the hash algorithm itself,
as it's easier to automatically infer a context with hashInit.

make the hmac API to not take a hash algorithm parameter, but infer the
algorithm used by the a type. add a hmacAlg, that just like hashInitAlg
take an explicit algorithm as parameter to infer the HMAC type. this function
is better when type inference is not able to automatically deduce the phantom.